+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Added a `--skipTypecheck` CLI option to skip typechecking before tests.
+
 ### Fixed
 - Fixed an issue where typechecking errors when using rspack would not recover and required restarting the server. #TINY-14130